django/with-ai-llm/1_wikipedia_chat_assistant_using_llamaindex/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
banks==2.1.2
beautifulsoup4==4.13.4
black==25.1.0
certifi==2025.6.15
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distlib==0.3.9
distro==1.9.0
Django==5.2.3
djangorestframework==3.16.0
filelock==3.18.0
filetype==1.2.0
flake8==7.3.0
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.12
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
llama-cloud==0.1.26
llama-cloud-services==0.6.34
llama-index==0.12.43
llama-index-agent-openai==0.4.11
llama-index-cli==0.4.3
llama-index-core==0.12.43
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.7.7
llama-index-instrumentation==0.2.0
llama-index-llms-openai==0.4.7
llama-index-multi-modal-llms-openai==0.5.1
llama-index-program-openai==0.3.2
llama-index-question-gen-openai==0.3.1
llama-index-readers-file==0.4.9
llama-index-readers-llama-parse==0.4.0
llama-index-readers-wikipedia==0.3.0
llama-index-workflows==0.2.2
llama-parse==0.6.34
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
multidict==6.5.1
mypy==1.16.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
nodeenv==1.9.1
numpy==2.3.1
openai==1.91.0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pre_commit==4.2.0
propcache==0.3.2
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
pypdf==5.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
wikipedia==1.4.0
wrapt==1.17.2
yarl==1.20.1
